Split large files up into smaller ones by functionality
